Fixed things and Added NewMatchEntry Page
Fixed iPhone X Navbar.
Known Issues/Problems that need fixing:
Not able to add "+" button in navBarUI for iPhone and a FAB for android seperately (looking into it) - This will be used to access the NewMatchTimerEntry in the future. Currently, I just have a button on the screen for that.