Added Incredible Caching Feature
It's finally here... Welcome caching to the NRG Scouting App.
Yes, its not there for JSON because JSON Has not been developed, so oh well, but its here for the timer, appstate, and the back and save buttons in match events
If your app craches (which it never should) or you close it on accident, your timer and match entry page should open right up.