Match Events Are Now Visually working
If you add things to your match by the match timer tab, you can update the list in the match events by clicking the UPDATE button. This currently shows the same icons for all types of events (needs to be fixed by @nchhaj189 ) and also shows the type of event (drop, pick, climb, etc.) and time (for which I have made a parser that converts millisecond int values to string time values (this could perhaps be its own class that returns values to the app for brevity purposes...)

Currently, there are two variables in the MatchEventsFormat Class. One formats for parsing (ParseMatchEvents, which is a Dictionary) and the other for viewing on the match events page (ParseUserEvents, which is a List because of indexing issues w/ dictionaries)

@nchhaj189, work on deleting events when someone clicks one match event twice (make sure that it deletes the corresponding match event too (e.g.: delete cube picked, delets the cube drop that goes w/ it). You can see the changes I have made to the parser... Also, work on saving each match's data into a big string with all match events and parameters, with a unique separator.

@shreystechtips will work on sorting the match events list by time so that no matter when events are added, they are added in the correct order by time.