Can now parse backwards (from list to match String)
The app can now parse backwards from the List version of match events so that when a user deletes an event, it dissapears from the Match Events page and it goes away from the match string where events are stored.

This action can be invoked by:
MatchEventsFormat.returnUserEventsAsString(yourListName);
Note that yourListName must follow the struct of <String,String> to be successfully parsed by the app.