New features and stability fixes
Can now save match events mid match if app crashes (Use the matchEventsFormat to parse the string that I get - this can be used to parse events when they are edited)
 - @nchhaj189 needs to work on match parameters saving and parsing if parameters exist

Debug buttons for match events are gone and the page now autoupdates

Started work on match events full string. The format is:
      -  &matchParameters*matchEvents (repeated)
   In this, the string can be split by the "&" character for each match and then by the "*" for events and parameters (which can be parsed by the EventsFormat parser class to obtain data put into the entry)

We still need to work on parsing the full string into the value for parameters and (and grab the teamname, matchnumber, and field side) put it into a list/dictionary and then use that for the listView in the Matches Page. Their onClick, it can reference that part of the Array and pass that data into the temp string and populate the Match Entry timer and parameters pages.

@nchhaj189  Need your help on this URGENTLY!!!